Add files via upload
The code is a set of SQL queries that analyze COVID-19 data from two tables, "CovidDeaths" and "CovidVaccinations". It selects and filters data based on different criteria such as country, continent, and population. It calculates metrics like death percentages, infection rates, and rolling people vaccinated. The code also creates temporary tables and a view to store the analyzed data